Move tests and sources into separate modules
