+++ .github/workflows/python-app.yml
@@ -37,5 +37,5 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with unittest
       run: |
-        python test_engines.py
-        python test_export.py
+        python -m unittest tests.test_engines
+        python -m unittest tests.test_export
