+++ youtubeanalyzer/defines.py
